prep for plugins -> background change in cypress 4
