+++ examples/blogs__codepen-demo/cypress/background/index.js
@@ -1,0 +1,18 @@
+// ***********************************************************
+// This example background/index.js can be used to load plugins
+// in the background process
+//
+// You can change the location of this file or turn off loading
+// the background file with the 'backgroundFile' configuration option.
+//
+// You can read more here:
+// https://on.cypress.io/background-process
+// ***********************************************************
+
+// This function is called when a project is opened or re-opened (e.g. due to
+// the project's config changing)
+
+module.exports = (on, config) => {
+  // `on` is used to hook into various events Cypress emits
+  // `config` is the resolved Cypress config
+}
