+++ examples/blogs__vue-vuex-rest/cypress/background/index.js
@@ -1,0 +1,54 @@
+// ***********************************************************
+// This example background/index.js can be used to load plugins
+// in the background process
+//
+// You can change the location of this file or turn off loading
+// the background file with the 'backgroundFile' configuration option.
+//
+// You can read more here:
+// https://on.cypress.io/background-process
+// ***********************************************************
+
+// This function is called when a project is opened or re-opened (e.g. due to
+// the project's config changing)
+
+/* eslint-disable no-console */
+/* global Promise */
+
+const fs = require('fs')
+const path = require('path')
+
+const findRecord = (title) => {
+  const dbFilename = path.join(__dirname, '..', '..', 'data.json')
+  const contents = JSON.parse(fs.readFileSync(dbFilename))
+  const todos = contents.todos
+  return todos.find((record) => record.title === title)
+}
+
+const hasRecordAsync = (title, ms) => {
+  const delay = 50
+  return new Promise((resolve, reject) => {
+    if (ms < 0) {
+      return reject(new Error(`Could not find record with title "${title}"`))
+    }
+    const found = findRecord(title)
+    if (found) {
+      return resolve(found)
+    }
+    setTimeout(() => {
+      hasRecordAsync(title, ms - delay).then(resolve, reject)
+    }, 50)
+  })
+}
+
+module.exports = (on, config) => {
+  // `on` is used to hook into various events Cypress emits
+  // `config` is the resolved Cypress config
+  on('task', {
+    hasSavedRecord (title, ms = 3000) {
+      console.log('looking for title "%s" in the database (time limit %dms)',
+        title, ms)
+      return hasRecordAsync(title, ms)
+    },
+  })
+}
