+++ examples/preprocessors__typescript-browserify/cypress/background/index.js
