+++ examples/preprocessors__typescript-webpack/cypress/background/index.js
